made registration by email
