+++ Kraina/Kraina/AppDelegate.swift
@@ -22,8 +22,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         if let key = Bundle.main.object(forInfoDictionaryKey: "GoogleMapsApiKey") as? String {
             GMSServices.provideAPIKey("\(key)")
         }
-        
-        
+        FirebaseApp.configure()
+
         return true
     }
