+++ Kraina/Kraina/ViewControllers/RegistrationViewController.swift
@@ -6,8 +6,12 @@
 //
 
 import UIKit
+import Firebase
 
 class RegistrationViewController: UIViewController {
+    
+    //MARK: - Создание переменных
+    var successfullLabel: (()->())?
     
     //MARK: - Создание элементов UI
     private lazy var registerView: UIView = {
@@ -24,47 +28,50 @@ class RegistrationViewController: UIViewController {
         return nameLabel
     }()
     
-   private lazy var nameTextField: UITextField = {
+    private lazy var emailTextField: UITextField = {
         let textField = UITextField()
         textField.translatesAutoresizingMaskIntoConstraints = false
-        textField.placeholder = "Введите Логин"
+        textField.placeholder = "Введите email"
         textField.keyboardType = UIKeyboardType.default
         textField.returnKeyType = UIReturnKeyType.done
         textField.autocorrectionType = UITextAutocorrectionType.no
         textField.font = UIFont.systemFont(ofSize: 13)
         textField.borderStyle = UITextField.BorderStyle.roundedRect
         textField.clearButtonMode = UITextField.ViewMode.whileEditing;
+        textField.keyboardType = .emailAddress
         textField.contentVerticalAlignment = UIControl.ContentVerticalAlignment.center
         return textField
     }()
     
     private lazy var passwordTextField: UITextField = {
-         let textField = UITextField()
-         textField.translatesAutoresizingMaskIntoConstraints = false
-         textField.placeholder = "Введите Пароль"
-         textField.keyboardType = UIKeyboardType.default
-         textField.returnKeyType = UIReturnKeyType.done
-         textField.autocorrectionType = UITextAutocorrectionType.no
-         textField.font = UIFont.systemFont(ofSize: 13)
-         textField.borderStyle = UITextField.BorderStyle.roundedRect
-         textField.clearButtonMode = UITextField.ViewMode.whileEditing;
-         textField.contentVerticalAlignment = UIControl.ContentVerticalAlignment.center
-         return textField
-     }()
+        let textField = UITextField()
+        textField.translatesAutoresizingMaskIntoConstraints = false
+        textField.placeholder = "Введите пароль"
+        textField.keyboardType = UIKeyboardType.default
+        textField.returnKeyType = UIReturnKeyType.done
+        textField.autocorrectionType = UITextAutocorrectionType.no
+        textField.font = UIFont.systemFont(ofSize: 13)
+        textField.borderStyle = UITextField.BorderStyle.roundedRect
+        textField.clearButtonMode = UITextField.ViewMode.whileEditing;
+        textField.isSecureTextEntry = true
+        textField.contentVerticalAlignment = UIControl.ContentVerticalAlignment.center
+        return textField
+    }()
     
     private lazy var confirmPasswordTextField: UITextField = {
-         let textField = UITextField()
-         textField.translatesAutoresizingMaskIntoConstraints = false
-         textField.placeholder = "Подтвердите пароль"
-         textField.keyboardType = UIKeyboardType.default
-         textField.returnKeyType = UIReturnKeyType.done
-         textField.autocorrectionType = UITextAutocorrectionType.no
-         textField.font = UIFont.systemFont(ofSize: 13)
-         textField.borderStyle = UITextField.BorderStyle.roundedRect
-         textField.clearButtonMode = UITextField.ViewMode.whileEditing;
-         textField.contentVerticalAlignment = UIControl.ContentVerticalAlignment.center
-         return textField
-     }()
+        let textField = UITextField()
+        textField.translatesAutoresizingMaskIntoConstraints = false
+        textField.placeholder = "Подтвердите пароль"
+        textField.keyboardType = UIKeyboardType.default
+        textField.returnKeyType = UIReturnKeyType.done
+        textField.autocorrectionType = UITextAutocorrectionType.no
+        textField.font = UIFont.systemFont(ofSize: 13)
+        textField.borderStyle = UITextField.BorderStyle.roundedRect
+        textField.clearButtonMode = UITextField.ViewMode.whileEditing;
+        textField.isSecureTextEntry = true
+        textField.contentVerticalAlignment = UIControl.ContentVerticalAlignment.center
+        return textField
+    }()
     
     private lazy var registerButton: UIButton = {
         let moveButton = UIButton()
@@ -75,7 +82,8 @@ class RegistrationViewController: UIViewController {
         moveButton.addTarget(self, action: #selector(self.registerButtonPressed), for: .touchUpInside)
         return moveButton
     }()
-
+    
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -85,13 +93,13 @@ class RegistrationViewController: UIViewController {
         //MARK: - Добавление элементов на экран
         view.addSubview(registerView)
         registerView.addSubview(titleRegisterLable)
-        registerView.addSubview(nameTextField)
+        registerView.addSubview(emailTextField)
         registerView.addSubview(passwordTextField)
         registerView.addSubview(confirmPasswordTextField)
         registerView.addSubview(registerButton)
         
         updateViewConstraints()
-
+        
     }
     
     //MARK: - Работа с констрейнтами
@@ -106,7 +114,7 @@ class RegistrationViewController: UIViewController {
             $0.top.equalToSuperview().inset(50)
         }
         
-        nameTextField.snp.makeConstraints {
+        emailTextField.snp.makeConstraints {
             $0.left.equalToSuperview().inset(20)
             $0.right.equalToSuperview().inset(20)
             $0.top.equalTo(titleRegisterLable).inset(50)
@@ -115,7 +123,7 @@ class RegistrationViewController: UIViewController {
         passwordTextField.snp.makeConstraints {
             $0.left.equalToSuperview().inset(20)
             $0.right.equalToSuperview().inset(20)
-            $0.top.equalTo(nameTextField).inset(50)
+            $0.top.equalTo(emailTextField).inset(50)
         }
         
         confirmPasswordTextField.snp.makeConstraints {
@@ -130,14 +138,32 @@ class RegistrationViewController: UIViewController {
             $0.top.equalTo(confirmPasswordTextField).inset(50)
             $0.height.equalTo(50)
         }
-
+        
+        
         super.updateViewConstraints()
     }
     
     //MARK: - Действие кнопки createAcc
     @objc private func registerButtonPressed() {
-        print("register")
+        if let email = emailTextField.text,
+           let passwordText = passwordTextField.text,
+           let confirmPasswordText = confirmPasswordTextField.text {
+            if isValidEmail(testStr: email), passwordText.count > 5, passwordText == confirmPasswordText {
+                Auth.auth().createUser(withEmail: email, password: passwordText) {[self] result, error in
+                    print(error)
+                    if let resultUnwrapped = result {
+                        print(resultUnwrapped.user.uid)
+                        
+                    } else {
+                        doErrorAlert(title: "Ошибка", message: "Возможно данный email уже зарегистрирован")
+                    }
+                }
+            } else {
+                doErrorAlert(title: "Ошибка", message: "Пожалуйста проверьте введенные данные")
+            }
+            print("register")
+        }
+        
+        
     }
-    
-
 }
